Re-structure settings views/components
