+++ src/renderer/components/app/index.ts
@@ -1,0 +1,1 @@
+export { AppSettings } from './AppSettings'
