--- src/renderer/components/wallet/settings/index.ts
@@ -1,1 +1,0 @@
-export { Settings } from './Settings'
