+++ src/renderer/views/wallet/WalletView.styles.ts
@@ -5,3 +5,6 @@ import { PoolActionsHistoryView } from './PoolActionsHistory'
 export const PoolActionsHistory = styled(PoolActionsHistoryView)`
   padding-top: 10px;
 `
+export const WalletSettingsWrapper = styled.div`
+  margin-top: 50px;
+`
